Move past_reimbursement_* at the top
Currently, if a reimbursement past the refund_time_constraint with a
preferred_reimbursement_type is performed : no type will be set since
:valid_preferred_reimbursement_type will return false and execution will
stop.

This results in "Unpaid amount is not zero" errors with no indication
that the reimbursement is past the allowed period.

By moving the past_reimbursement_type_period check at the top, we can
properly handle this case and correctly sets the
expired_reimbursement_type.

- Update the spec to reflect the new priority