Remove :valid_preferred_reimbursement_type
Previously, valid_preferred_reimbursement would validate if the
reimbursement was past the refund_constraint_time.

This is no longer a requirement as the refund_time_constraint validation
has been moved outside of this context.

Since this was only operation performed by the method, it can be safely
removed from ReimbursementTypeEngine and ReimbursementTypeValidator