Allow generic adjustments creation in admin
Currently the adjustment is initialized as an Order adjustment. Using
the base model instead of the parent allows us to create both LineItem
and Order adjustments.

The downside is that in memory Order representation can be stale as we
create an adjustment directly instead of building the association but in
this scenario there should no side effect as there is a hard
reload/update callback to ensure totals are correctly updated.